Revert: CI: add a build with sources of librmn
It caused SSM packages to be created using librmn sources, which are a at a
specific commit in rmn submodule, instead of using the current tagged
version of librmn. It also caused the installation of librmn and App
libraries (at that specific commit) with xrec.

A better way of testing xrec for external users will be added later.